Added pass change

+++ Backend/PharmacyProject/src/main/java/com/team11/PharmacyProject/offer/OfferService.java
@@ -8,5 +8,5 @@ public interface OfferService {
 
     List<Offer> findOffersByOrderId(long orderId);
 
-    int calculateExpenses(long start, long end, long pharmacyId);
+    double calculateExpenses(long start, long end, long pharmacyId);
 }
